CWMS-1256 update byte total to long to avoid overflow

+++ cumulus-client/src/main/java/mil/army/usace/hec/cumulus/client/controllers/CumulusDssFileDownloadListener.java
@@ -5,5 +5,5 @@ import mil.army.usace.hec.cumulus.client.model.Download;
 
 public interface CumulusDssFileDownloadListener {
 
-    void bytesRead(Download downloadData, int currentBytesRead, int totalBytesRead, Duration elapsedTime);
+    void bytesRead(Download downloadData, int currentBytesRead, long totalBytesRead, Duration elapsedTime);
 }
